OK, the path attribute is actually a filename.
getPath() is missleading then, too. However it is a path for the first
directories to be examined and a relative path to the parent directory
(effectively a filename) for all children.

Hence, the win32 module could work for now. Missing, of course, is a
more advanced directory comparison that takes the directory's attributes
into account.